core: Reorganize header files
Change the header organization:
 - Move legacy API into a separate header
 - Add header with new SUIT API
 - Move internal structures to suit_processor header file
